requirements.txt
tensorflow>=2.12.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
